Tons of changes
I made a lot of chanegs on they way that the mvoemon of the sprites works, first you have a preview in the left side there you can decide its color type.
When you clicked on a sprite and pass your mouse trhough the baord a "shadow" to where the sprite will be put will appear. Right now it doesnt detect when the sprites had rotated nor does it detecet "collision". Both features shoul be implemented tomorrow.